[bugfix] fix bug of writer link into a cycle.
notify follower before new-leader, the actual new-leader already point to
the last-follwer, if last-follower can into next loop(I use the address
of writer in stack to mark the unique writer), then last follower can
point to the actual new-leader. And if I use older==last-follower to distinguish
the new-leader, I will find a wrong new-leader, also leave the other
writer which is in a circle forever.